parser + lexer working
